[510682] Adds check for Dynamic Web facet before version check
In the WebAnnotationReader.analyzeCompilationUnit function, there is a
check to see what version of the Dynamic Web facet is installed in a
project without checking to see if the facet is even installed. A simple
solution is to add a check to see if the facet is installed, before
checking the version, like what was done in
Web25MergedModelProvider.createWebMerger.

Signed-off-by: John Collier <John.J.Collier@ibm.com>