Correct getInterpretedSpecVersion(..) in J2EEArtifactImportDataModelProvider to use the highest version when nothing else is known (#8)
Assuming the highest version is the expected behaviour according to the JavaDoc. It is also the the behaviour that was present in 487b0b36eaa624f1723a88aab771dc2f8cee51f9 before the changes for Jakarta EE 10 in 9788800380e7f38cfa7c657d11ddf1b66ac587cd . The changes to support Jakarta EE 10 clamped the version at EE 8.